LEARNER-506, fixed comments, code improvements
