+++ Test/NSAttributedString+FormattingTests.swift
@@ -19,14 +19,4 @@ class NSAttributedString_FormattingTests: XCTestCase {
         let result = styled(name)
         XCTAssertEqual(result.string, "Hello someone")
     }
-    
-    func testDecodingHTMLEntities() {
-        XCTAssertEqual("what is it that you&amp;#39;re really looking for when you conduct a user test?".decodingHTMLEntities, "what is it that you're really looking for when you conduct a user test?")
-        XCTAssertEqual("what is it that you&#39;re really looking for when you conduct a user test?".decodingHTMLEntities, "what is it that you're really looking for when you conduct a user test?")
-        XCTAssertEqual("This is awesome &hearts;".decodingHTMLEntities, "This is awesome ♥")
-        XCTAssertEqual("This is awesome &amp;&hearts;".decodingHTMLEntities, "This is awesome &♥")
-        XCTAssertEqual("we&#39;ve done with option one & two".decodingHTMLEntities, "we've done with option one & two")
-        XCTAssertEqual("we&amp;#39;ve done with option one & two".decodingHTMLEntities, "we've done with option one & two")
-        XCTAssertEqual("we&amp;#39;ve done with option &one; & &two;".decodingHTMLEntities, "we've done with option &one; & &two;")
-    }
 }
